requirements.txt
aiohttp==3.7.3
altgraph==0.17
amqp==5.0.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.8.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
cached-property==1.5.1
camelot-py==0.8.2
celery==5.0.5
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cmake==3.18.4.post1
colorama==0.4.3
command-not-found==0.3
configparser==3.5.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cupshelpers==1.0
cycler==0.10.0
dateparser==1.0.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.44.0
django-auth-token==0.0.23
django-ckeditor==6.0.0
django-colorfield==0.3.2
django-core==1.4.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-ipware==3.0.2
django-js-asset==1.2.2
django-loginas==0.3.9
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
django-silk==4.1.0
django-summernote==0.8.11.6
django-token==0.1.3
djangorestframework==3.12.4
dlib==19.21.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
docx==0.2.4
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
et-xmlfile==1.0.1
evdev==1.4.0
eventkit==0.8.6
excalibur-py==0.4.3
face-recognition==1.3.0
face-recognition-models==0.3.0
facebook-sdk==3.1.0
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
getmac==0.8.2
gevent==21.1.2
gevent-websocket==0.10.1
gprof2dot==2019.11.30
greenlet==1.0.0
gunicorn==20.0.4
httplib2==0.14.0
huepy==1.2.1
hyperlink==21.0.0
ib-insync==0.9.64
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3
import-string==0.1.0
importlib-metadata==1.5.0
incremental==17.5.0
iniconfig==1.1.1
instabot==0.117.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kazam==1.4.5
keepa==1.1.0
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.3.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
meld==3.20.2
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
moviepy==1.0.3
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.1.6
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pbr==5.5.1
pdfminer.six==20201018
pendulum==2.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
plyer==2.0.0
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
Pygments==2.7.3
PyGObject==3.36.0
PyHamcrest==2.0.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynput==1.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
pytesseract==0.3.7
pytest==6.2.2
python-apt==2.0.0+ubuntu0.20.4.3
python-binance==0.7.9
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-docx==0.8.10
python-twitter==3.5
python-xlib==0.29
python3-linkedin==1.0.2
python3-openid==3.2.0
pyttk==0.3.2
pytz==2019.3
pytzdata==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.12.1
schedule==1.0.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simpleaudio==1.0.4
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smtp==0.1.0
sortedcontainers==2.3.0
soupsieve==2.1
SQLAlchemy==1.3.22
sqlparse==0.2.4
stevedore==3.3.0
supervisor==4.1.0
systemd-python==234
tabula-py==2.2.0
terminado==0.9.1
testpath==0.4.4
texttable==1.6.2
tkcalendar==1.6.1
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
tweepy==3.10.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.3
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.16.1
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
yarl==1.6.3
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.2.0
